feat(App, Lotto.js) : fix functions
